Move `Layout*` components into a folder
There are a whole bunch of layout related components. This just isolates
them into a nice little directory of their own.
