+++ App.js
@@ -6,8 +6,8 @@ import React from 'react';
 // eslint-disable-next-line import/default
 import Expo from 'expo';
 
-import LayoutManager from './src/components/LayoutManager';
-import LayoutConfig from './src/components/LayoutConfig';
+import LayoutManager from './src/components/layout/LayoutManager';
+import LayoutConfig from './src/components/layout/LayoutConfig';
 import ServerLink from './src/components/ServerLink';
 import * as colors from './src/constants/colors';
 import Loading from './src/components/Loading';
