+++ src/components/layout/LayoutConfig.js
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import R from 'ramda';
 
-import * as colors from '../constants/colors';
-import * as actions from '../actions/layout';
-import { selector } from '../utils/redux';
+import * as colors from '../../constants/colors';
+import * as actions from '../../actions/layout';
+import { selector } from '../../utils/redux';
 
 const Container = styled.View`
   background-color: ${colors.groups.bg};
