+++ src/components/layout/LayoutOption.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import R from 'ramda';
 
-import * as colors from '../constants/colors';
+import * as colors from '../../constants/colors';
 
 export const Container = styled.View`
   background-color: ${colors.groups.bg};
