+++ src/components/layout/__tests__/LayoutManager.test.js
